Increased line limit and added console warning when line limit is reached
I was trying to implement the warning in the 'onTextViewerTextChanged()' method but this method is recursively being called. I assume it has to do with the fact that every line counts as a change in the textViewer so it'll emit the signal for every line. Without getting outside of scope I couldn't figure it out how to fix it. So I had to implement the warning in the loadTextFile() method instead.